requirements.txt
pyyaml
crewai-tools
crewai>=0.177.0
# crewai>=0.36.0
langchain>=0.2.10
openai>=1.40.0
python-dotenv>=1.0.1
GitPython>=3.1.43
PyGithub>=2.3.0
pandas>=2.2.2
tqdm>=4.66.4
